Conditional at Launch
Previously, a binding was required at build time in order to contribute the
layers for the Application Insights components.  We've started to think that
requiring a binding during build is too onerous a restriction and instead
environment variables should be used to indicate that a dependency should be
contributed. The knock-on effect of this change is that we have to move a lot
more of the conditional logic out to launch time including ensuring that if
the binding doesn't exist _nothing_ is contributed.  This change makes those
updates to the buildpack's behavior.

In additional, this includes a fix where the provided build plan wasn't
exhaustive for combinations that a user need when using the buildpack.

Signed-off-by: Ben Hale <bhale@vmware.com>